refactor: Extract sqlalchemy compat into its own module.
